tests: Move TestAddressee to test_message_send.py.
